Refactor in progress
