+++ src/utils/api.js
@@ -1,0 +1,9 @@
+class Api {
+  constructor(options) {
+    this._options = options;
+  }
+
+  getInitialCards() {}
+}
+
+export default Api;
